+++ CHANGELOG.md
@@ -21,7 +21,10 @@ and start a new "In Progress" section above it.
 
 ## In progress
 
+## 0.107.0
+
 - `evaluate_process_from_url`: drop support for URL guessing from folder-like URL ([#297)](https://github.com/Open-EO/openeo-python-driver/issues/297))
+- `evaluate_process_from_url`: align with new (and experimental) "remote-udp" extension ([#297)](https://github.com/Open-EO/openeo-python-driver/issues/297))
 
 ## 0.106.0